requirements.txt
Django
django-extensions
django-filter
djangorestframework
Markdown
